backend/requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
Flask-Migrate
Flask-CORS
Flask-Login
Flask-WTF
requests
tweepy
pandas
numpy
authlib
werkzeug
Pillow